mcp-server-omni/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-server-omni"
version = "0.2.2"
description = "A Model Context Protocol server for Omni ERP systems"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Andrey Ivanov", email = "ivnv.xd@gmail.com" }
]
maintainers = [
    { name = "Andrey Ivanov", email = "ivnv.xd@gmail.com" }
]
license = { text = "MPL-2.0" }
keywords = ["mcp", "omni", "erp", "model-context-protocol", "ai", "llm", "claude", "anthropic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mcp>=1.9.4",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/ivnvxd/mcp-server-omni"
Documentation = "https://github.com/ivnvxd/mcp-server-omni#readme"
Repository = "https://github.com/ivnvxd/mcp-server-omni"
Issues = "https://github.com/ivnvxd/mcp-server-omni/issues"
Changelog = "https://github.com/ivnvxd/mcp-server-omni/releases"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "mypy>=1.16.0",
    "ruff>=0.11.12",
    "black>=25.1.0",
    "types-requests>=2.32.0",
    "requests>=2.32.3",
]

[project.scripts]
mcp-server-omni = "mcp_server_omni.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server_omni"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["."]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.coverage.run]
source = ["mcp_server_omni"]
omit = ["tests/*", "*/__pycache__/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_any_unimported = false
ignore_missing_imports = true
no_implicit_optional = false
check_untyped_defs = false
strict_optional = false
allow_untyped_calls = true
allow_incomplete_defs = true
allow_untyped_globals = true
allow_redefinition = true
no_warn_no_return = true
warn_unreachable = false
# Disable specific error codes that are problematic
disable_error_code = ["str-bytes-safe", "union-attr", "arg-type", "return-value", "assignment", "list-item", "var-annotated"]
exclude = [
    "venv/",
    ".venv/",
    "build/",
    "dist/",
]

[[tool.mypy.overrides]]
module = "mcp.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
ignore_errors = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "Q"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310"]

[dependency-groups]
dev = [
    "mypy>=1.16.0",
]
